startlist update - 3 new buttons

+++ src/main/java/fei/stuba/bp/rigo/preteky/models/sql/ResultStartList.java
@@ -26,7 +26,7 @@ public class ResultStartList implements Serializable {
     private Discipline discipline;
 
     @Column(name = "start_performance")
-    private Integer startPerformance;
+    private Double startPerformance;
 
     @Column(name = "result_performance")
     private Integer resultPerformance;
